requirements.txt
mysql-connector-python==8.0.22
beautifulsoup4==4.9.3
requests==2.24.0
html2text==2020.1.16